+++ config/environments/test.rb
@@ -55,9 +55,6 @@ module OptimizerWrapper
   OptimizerLogger.with_datetime = true
   # OptimizerLogger.caller_location = nil => nil is default
 
-  I18n.available_locales = [:en, :fr]
-  I18n.default_locale = :en
-
   @@c = {
     product_title: 'Optimizers API',
     product_contact_email: 'tech@mapotempo.com',