+++ models/vrp.rb
@@ -79,9 +79,11 @@ module Models
     field :restitution_geometry, default: []
     field :restitution_intermediate_solutions, default: true
     field :restitution_csv, default: false
+    field :restitution_use_deprecated_csv_headers, default: false
     field :restitution_allow_empty_result, default: false
 
     field :schedule_range_indices, default: nil # extends schedule_range_date
+    field :schedule_start_date, default: nil
     field :schedule_unavailable_days, default: Set[] # extends unavailable_date and schedule_unavailable_indices
     field :schedule_months_indices, default: []
 
@@ -435,6 +437,7 @@ module Models
         start: start_index,
         end: end_index
       }
+      hash[:configuration][:schedule][:start_date] = hash[:configuration][:schedule][:range_date][:start]
       hash[:configuration][:schedule].delete(:range_date)
 
       hash
@@ -460,6 +463,7 @@ module Models
       self.restitution_geometry = restitution[:geometry]
       self.restitution_intermediate_solutions = restitution[:intermediate_solutions]
       self.restitution_csv = restitution[:csv]
+      self.restitution_use_deprecated_csv_headers = restitution[:use_deprecated_csv_headers]
       self.restitution_allow_empty_result = restitution[:allow_empty_result]
     end
 
@@ -504,6 +508,7 @@ module Models
 
     def schedule=(schedule)
       self.schedule_range_indices = schedule[:range_indices]
+      self.schedule_start_date = schedule[:start_date]
       self.schedule_unavailable_days = schedule[:unavailable_days]
       self.schedule_months_indices = schedule[:months_indices]
     end